feat: separate on layers according FSD-architecture
