+++ src/entities/content-top-left/ContentTopLeft.module.scss
