+++ src/entities/content-top-right/ContentTopRight.module.scss
