+++ src/entities/content-top-right/ContentTopRight.tsx
