+++ src/entities/index.ts
@@ -1,0 +1,2 @@
+export { ContentTopLeft } from './content-top-left/ContentTopLeft';
+export { ContentTopRight } from './content-top-right/ContentTopRight';
