+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export { Login } from './login/Login';
+export { Main } from './main/Main';
+export { Register } from './register/Register';
