+++ src/widgets/auth-left-block/AuthLeftBlock.module.scss
