+++ src/widgets/content-top/ContentTop.module.scss
