+++ src/widgets/content-tracks/ContentTracks.module.scss
