+++ src/widgets/index.ts
@@ -1,8 +1,7 @@
-export { LoginForm } from './login-form/LoginForm';
 export { AuthLeftBlock } from './auth-left-block/AuthLeftBlock';
+export { LoginForm } from './login-form/LoginForm';
+export { RegisterForm } from './register-form/RegisterForm';
 export { ContentTop } from './content-top/ContentTop';
-export { ContentTopLeft } from './content-top-left/ContentTopLeft';
-export { ContentTopRight } from './content-top-right/ContentTopRight';
 export { MainHeader } from './main-header/MainHeader';
 export { MainLogo } from './main-logo/MainLogo';
 export { MainMenu } from './main-menu/MainMenu';
