+++ src/widgets/main-header/MainHeader.module.scss
