+++ src/widgets/main-logo/MainLogo.module.scss
