+++ src/widgets/main-logo/MainLogo.tsx
