+++ src/widgets/main-menu/MainMenu.module.scss
