+++ src/widgets/register-form/RegisterForm.module.scss
